Copied mild tutorial updaes to new master

+++ Five Stars in the Dark/Assets/Scripts/Level/Scoring/TrackErrors.cs
@@ -20,7 +20,7 @@ public class TrackErrors : MonoBehaviour
     }
 
     //because errors is static, it needs a method to access
-    public static float getErrors()
+    public static int getErrors()
     {
         return errors;
     }
